requirements.txt
Flask
Flask-SQLAlchemy
SQLAlchemy
Flask-Migrate
Flask-WTF
WTForms
python-dotenv
Flask-Caching
Flask-Login
langchain
langchain-community
langchain-core
langchain-groq
langchain-openai
langchain-google-genai
faiss-cpu
pypdf
groq
openai
werkzeug
requests
google-generativeai
langchain-text-splitters
pydantic
pydantic-core
pydantic-settings
numpy
tqdm
tenacity
PyYAML
regex